requirements.txt
# Core CrewAI Framework
crewai[tools]>=0.41.1
crewai-tools>=0.4.26

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Redis
redis>=5.0.1
aioredis>=2.0.1

# Async Support
asyncio>=3.4.3
aiofiles>=23.2.1
aiohttp>=3.9.1

# Data Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Environment and Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# CLI and Interface
click>=8.1.7
rich>=13.7.0
prompt-toolkit>=3.0.43

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.2

# Logging and Monitoring
loguru>=0.7.2
structlog>=24.1.0

# Utilities
httpx>=0.26.0
tenacity>=8.2.3
python-dateutil>=2.8.2

# Type Checking
types-redis>=4.6.0.11
types-aiofiles>=23.2.0.0
types-pyyaml>=6.0.12.12